+++ src/service/ConfigService.ts
@@ -268,7 +268,8 @@ export class ConfigService {
         const copyFrom = join(this.root, 'config', 'node');
         const name = account.name;
 
-        const outputFolder = BootstrapUtils.getTargetNodesFolder(this.params.target, false, name, 'userconfig');
+        const serverConfig = BootstrapUtils.getTargetNodesFolder(this.params.target, false, name, 'server-config');
+        const brokerConfig = BootstrapUtils.getTargetNodesFolder(this.params.target, false, name, 'broker-config');
         const nodePreset = (presetData.nodes || [])[index];
 
         const nodePrivateKey = await CommandUtils.resolvePrivateKey(
@@ -324,11 +325,28 @@ export class ConfigService {
             await BootstrapUtils.generateConfiguration(templateContext, copyFrom, agentConfig, []);
         }
 
-        await BootstrapUtils.generateConfiguration(templateContext, copyFrom, outputFolder, excludeFiles);
+        const serverRecoveryConfig = {
+            addressextractionRecovery: false,
+            mongoRecovery: false,
+            zeromqRecovery: false,
+            filespoolingRecovery: true,
+            hashcacheRecovery: true,
+        };
+
+        const brokerRecoveryConfig = {
+            addressextractionRecovery: true,
+            mongoRecovery: true,
+            zeromqRecovery: true,
+            filespoolingRecovery: false,
+            hashcacheRecovery: true,
+        };
+
+        logger.info(`Generating ${name} server configuration`);
+        await BootstrapUtils.generateConfiguration({ ...serverRecoveryConfig, ...templateContext }, copyFrom, serverConfig, excludeFiles);
         await this.generateP2PFile(
             presetData,
             addresses,
-            outputFolder,
+            serverConfig,
             NodeType.PEER_NODE,
             (nodePresetData) => nodePresetData.harvesting,
             'peers-p2p.json',
@@ -336,11 +354,38 @@ export class ConfigService {
         await this.generateP2PFile(
             presetData,
             addresses,
-            outputFolder,
+            serverConfig,
             NodeType.API_NODE,
             (nodePresetData) => nodePresetData.api,
             'peers-api.json',
         );
+
+        if (nodePreset.brokerName) {
+            logger.info(`Generating ${nodePreset.brokerName} broker configuration`);
+            await BootstrapUtils.generateConfiguration(
+                { ...brokerRecoveryConfig, ...templateContext },
+                copyFrom,
+                brokerConfig,
+                excludeFiles,
+            );
+            await this.generateP2PFile(
+                presetData,
+                addresses,
+                brokerConfig,
+                NodeType.PEER_NODE,
+                (nodePresetData) => nodePresetData.harvesting,
+                'peers-p2p.json',
+            );
+            await this.generateP2PFile(
+                presetData,
+                addresses,
+                brokerConfig,
+                NodeType.API_NODE,
+                (nodePresetData) => nodePresetData.api,
+                'peers-api.json',
+            );
+        }
+
         await new VotingService(this.params).run(presetData, account, nodePreset);
     }
 
@@ -388,7 +433,7 @@ export class ConfigService {
         const transactionsDirectory = join(nemesisWorkingDir, presetData.nemesis.transactionsDirectory || presetData.transactionsDirectory);
         await BootstrapUtils.mkdir(transactionsDirectory);
         const copyFrom = join(this.root, `config`, `nemesis`);
-        const moveTo = join(nemesisWorkingDir, `userconfig`);
+        const moveTo = join(nemesisWorkingDir, `server-config`);
         const templateContext = { ...(presetData as any), addresses };
         await Promise.all(
             (addresses.nodes || []).filter((n) => n.vrf).map((n) => this.createVrfTransaction(transactionsDirectory, presetData, n)),
@@ -557,7 +602,7 @@ export class ConfigService {
                     this.params.target,
                     false,
                     gatewayPreset.apiNodeName,
-                    'userconfig',
+                    'server-config',
                     'resources',
                 );
                 const apiNodeCertFolder = BootstrapUtils.getTargetNodesFolder(this.params.target, false, gatewayPreset.apiNodeName, 'cert');
@@ -632,8 +677,15 @@ export class ConfigService {
     private cleanUpConfiguration(presetData: ConfigPreset) {
         const target = this.params.target;
         (presetData.nodes || []).forEach(({ name }) => {
-            const serverConfigFolder = BootstrapUtils.getTargetNodesFolder(target, false, name, 'userconfig');
+            const serverConfigFolder = BootstrapUtils.getTargetNodesFolder(target, false, name, 'server-config');
             BootstrapUtils.deleteFolder(serverConfigFolder);
+
+            const brokerConfigFolder = BootstrapUtils.getTargetNodesFolder(target, false, name, 'broker-config');
+            BootstrapUtils.deleteFolder(brokerConfigFolder);
+
+            // Remove old user configs when upgrading.
+            const userConfigFolder = BootstrapUtils.getTargetNodesFolder(target, false, name, 'userconfig');
+            BootstrapUtils.deleteFolder(userConfigFolder);
 
             const seedFolder = BootstrapUtils.getTargetNodesFolder(target, false, name, 'seed');
             BootstrapUtils.deleteFolder(seedFolder);